fix(js): react app compatibility fix
