--- ts/src/static_dependencies/qs/index.cjs
@@ -1,11 +1,0 @@
-'use strict';
-
-var stringify = require('./stringify.cjs');
-var parse = require('./parse.cjs');
-var formats = require('./formats.cjs');
-
-module.exports = {
-    formats: formats,
-    parse: parse,
-    stringify: stringify
-};
