+++ ts/src/static_dependencies/qs/stringify.js
@@ -1,7 +1,5 @@
-'use strict';
-
-var utils = require('./utils.cjs');
-var formats = require('./formats.cjs');
+import * as utils from './utils.js';
+import defaultFormat, { formatters } from './formats.js';
 var has = Object.prototype.hasOwnProperty;
 
 var arrayPrefixGenerators = {
@@ -34,7 +32,7 @@ var defaults = {
     encode: true,
     encoder: utils.encode,
     encodeValuesOnly: false,
-    formatter: formats.formatters[formats['default']],
+    formatter: formatters[defaultFormat],
     // deprecated
     indices: false,
     serializeDate: function serializeDate(date) { // eslint-disable-line func-name-matching
@@ -157,14 +155,14 @@ var normalizeStringifyOptions = function normalizeStringifyOptions(opts) {
         throw new TypeError('The charset option must be either utf-8, iso-8859-1, or undefined');
     }
 
-    var format = formats['default'];
+    var format = defaultFormat;
     if (typeof opts.format !== 'undefined') {
-        if (!has.call(formats.formatters, opts.format)) {
+        if (!has.call(formatters, opts.format)) {
             throw new TypeError('Unknown format option provided.');
         }
         format = opts.format;
     }
-    var formatter = formats.formatters[format];
+    var formatter = formatters[format];
 
     var filter = defaults.filter;
     if (typeof opts.filter === 'function' || isArray(opts.filter)) {
@@ -189,7 +187,7 @@ var normalizeStringifyOptions = function normalizeStringifyOptions(opts) {
     };
 };
 
-module.exports = function (object, opts) {
+export default function (object, opts) {
     var obj = object;
     var options = normalizeStringifyOptions(opts);
 
@@ -266,4 +264,4 @@ module.exports = function (object, opts) {
     }
 
     return joined.length > 0 ? prefix + joined : '';
-};
+}
